IPIV is int vector. Not double.

+++ tests/LU/dgesv.c
@@ -579,7 +579,7 @@ static void create_datatypes(void)
     MPI_Type_commit(&LITTLE_L);
 
     /* IPIV is a contiguous of size N */
-    MPI_Type_contiguous(N, MPI_DOUBLE, &PIVOT_VECT);
+    MPI_Type_contiguous(N, MPI_INT, &PIVOT_VECT);
     MPI_Type_set_name(PIVOT_VECT, "Pivot vector");
     MPI_Type_commit(&PIVOT_VECT);
